refactor: refine code and add some doc
