+++ pkg/p2p/configure/configure.go
@@ -14,17 +14,52 @@
    limitations under the License.
 */
 
-package p2p
+package configure
 
 import (
-	"crypto/tls"
 	"fmt"
-	"net/http"
+
+	"github.com/alibaba/accelerated-container-image/pkg/p2p/util"
 
 	log "github.com/sirupsen/logrus"
 	"github.com/spf13/viper"
 )
 
+// DeployConfig is server config
+type DeployConfig struct {
+	LogLevel       string
+	RunMode        string
+	RootList       []string
+	NodeIP         string
+	DetectAddr     string
+	ServeBySSL     bool
+	Port           int
+	CertConfig     DeployCertConfig
+	CacheConfig    DeployCacheConfig
+	PrefetchConfig DeployPrefetchConfig
+}
+
+type DeployCertConfig struct {
+	CertEnable   bool
+	GenerateCert bool
+	CertPath     string
+	KeyPath      string
+}
+
+type DeployCacheConfig struct {
+	FileCacheEnable bool
+	FileCacheSize   int64
+	FileCachePath   string
+	MemCacheEnable  bool
+	MemCacheSize    int64
+}
+
+type DeployPrefetchConfig struct {
+	PrefetchEnable bool
+	PrefetchThread int
+}
+
+// CheckConfig used to check config and fix some configuration
 func CheckConfig(config *DeployConfig) {
 	// run mode
 	if config.RunMode != "root" && config.RunMode != "agent" {
@@ -44,12 +79,16 @@ func CheckConfig(config *DeployConfig) {
 	}
 	// certificate
 	if config.CertConfig.CertEnable {
-		config.CertConfig.CertPath = GetRealPath(config.CertConfig.CertPath)
-		config.CertConfig.KeyPath = GetRealPath(config.CertConfig.KeyPath)
+		config.CertConfig.CertPath = util.GetRealPath(config.CertConfig.CertPath)
+		config.CertConfig.KeyPath = util.GetRealPath(config.CertConfig.KeyPath)
+	}
+	// cache media
+	if config.CacheConfig.FileCacheEnable {
+		config.CacheConfig.FileCachePath = util.GetRealPath(config.CacheConfig.FileCachePath)
 	}
 	// nodeIP
 	if config.NodeIP == "" {
-		config.NodeIP = GetOutboundIP(config.DetectAddr).String()
+		config.NodeIP = util.GetOutboundIP(config.DetectAddr).String()
 	}
 	// log
 	level, err := log.ParseLevel(config.LogLevel)
@@ -67,6 +106,7 @@ func CheckConfig(config *DeployConfig) {
 	}
 }
 
+// InitConfig used to load config from disk
 func InitConfig(cfgFile string) *DeployConfig {
 	var config = &DeployConfig{}
 	viper.SetConfigFile(cfgFile)
@@ -79,50 +119,4 @@ func InitConfig(cfgFile string) *DeployConfig {
 	}
 	log.Println("Using config file:", viper.ConfigFileUsed())
 	return config
-}
-
-func Execute(config *DeployConfig, isRun bool) *http.Server {
-	switch config.RunMode {
-	case "root":
-		log.Info("Run on P2P Root")
-	case "agent":
-		log.Info("Run on P2P Agent")
-	}
-	cache := NewCachePool(&CacheConfig{
-		MaxEntry:   config.CacheConfig.MemCacheSize,
-		CacheSize:  config.CacheConfig.FileCacheSize,
-		CacheMedia: config.CacheConfig.FileCachePath,
-	})
-	hp := NewHostPicker(config.RootList, cache)
-	fs := NewP2PFS(&FSConfig{
-		CachePool:       cache,
-		HostPicker:      hp,
-		APIKey:          "dadip2p",
-		PrefetchWorkers: config.PrefetchConfig.PrefetchThread,
-	})
-	var myAddr string
-	if !config.ServeBySSL {
-		myAddr = fmt.Sprintf("http://%s:%d", config.NodeIP, config.Port)
-	} else {
-		myAddr = fmt.Sprintf("https://%s:%d", config.NodeIP, config.Port)
-	}
-	serverHandler := NewP2PServer(&ServerConfig{
-		MyAddr:     myAddr,
-		Fs:         fs,
-		APIKey:     "dadip2p",
-		ProxyHTTPS: config.CertConfig.CertEnable,
-	})
-	var cert []tls.Certificate
-	if config.CertConfig.CertEnable {
-		cert = []tls.Certificate{*GetRootCA(config.CertConfig.CertPath, config.CertConfig.KeyPath, config.CertConfig.GenerateCert)}
-	}
-	server := &http.Server{
-		Addr:      fmt.Sprintf(":%d", config.Port),
-		Handler:   serverHandler,
-		TLSConfig: &tls.Config{Certificates: cert},
-	}
-	if isRun {
-		panic(server.ListenAndServe())
-	}
-	return server
 }
