+++ pkg/p2p/test/cache_test.go
@@ -21,19 +21,22 @@ import (
 	"runtime"
 	"testing"
 
-	"github.com/alibaba/accelerated-container-image/pkg/p2p"
+	"github.com/alibaba/accelerated-container-image/pkg/p2p/cache"
+	"github.com/alibaba/accelerated-container-image/pkg/p2p/fs"
+	"github.com/alibaba/accelerated-container-image/pkg/p2p/util"
+
 	"github.com/stretchr/testify/assert"
 )
 
-func testCacheGetOrRefillHelper(t *testing.T, config *p2p.CacheConfig) {
+func testCacheGetOrRefillHelper(t *testing.T, config *cache.Config) {
 	t.Helper()
 	Assert := assert.New(t)
-	c := p2p.NewCachePool(config)
+	c := cache.NewCachePool(config)
 	for i := 0; i < 100; i++ {
-		fileName := p2p.GetRandomString(10)
+		fileName := util.GetRandomString(10)
 		fileContent := []byte(getData())
 		for j := 0; j < 10; j++ {
-			for seg := range p2p.NewRangeSplit(0, 128*1024, int64(len(fileContent)), int64(len(fileContent))).AllParts() {
+			for seg := range fs.NewRangeSplit(0, 128*1024, int64(len(fileContent)), int64(len(fileContent))).AllParts() {
 				wg.Add(1)
 				go func(offset int64, size int) {
 					defer wg.Done()
@@ -42,7 +45,7 @@ func testCacheGetOrRefillHelper(t *testing.T, config *p2p.CacheConfig) {
 					})
 					if Assert.Equal(nil, err) && Assert.Equal(size, len(res)) {
 						expected := fileContent[offset : offset+int64(size)]
-						checkLen := p2p.Min(100, size)
+						checkLen := fs.Min(100, size)
 						Assert.Equal(expected[:checkLen], res[:checkLen])
 						Assert.Equal(expected[len(res)-checkLen:], res[len(res)-checkLen:])
 					}
@@ -55,23 +58,23 @@ func testCacheGetOrRefillHelper(t *testing.T, config *p2p.CacheConfig) {
 }
 
 func TestCacheGetOrRefill(t *testing.T) {
-	testCacheGetOrRefillHelper(t, &p2p.CacheConfig{CacheSize: 100 * 1024 * 1024, MaxEntry: 0, CacheMedia: media})
-	testCacheGetOrRefillHelper(t, &p2p.CacheConfig{CacheSize: 90 * 1024 * 1024, MaxEntry: 0, CacheMedia: media})
-	testCacheGetOrRefillHelper(t, &p2p.CacheConfig{CacheSize: 10 * 1024 * 1024, MaxEntry: 0, CacheMedia: media})
-	testCacheGetOrRefillHelper(t, &p2p.CacheConfig{CacheSize: 1 * 1024 * 1024, MaxEntry: 0, CacheMedia: media})
-	testCacheGetOrRefillHelper(t, &p2p.CacheConfig{CacheSize: 0, MaxEntry: 0, CacheMedia: media})
+	testCacheGetOrRefillHelper(t, &cache.Config{CacheSize: 100 * 1024 * 1024, MaxEntry: 0, CacheMedia: media})
+	testCacheGetOrRefillHelper(t, &cache.Config{CacheSize: 90 * 1024 * 1024, MaxEntry: 0, CacheMedia: media})
+	testCacheGetOrRefillHelper(t, &cache.Config{CacheSize: 10 * 1024 * 1024, MaxEntry: 0, CacheMedia: media})
+	testCacheGetOrRefillHelper(t, &cache.Config{CacheSize: 1 * 1024 * 1024, MaxEntry: 0, CacheMedia: media})
+	testCacheGetOrRefillHelper(t, &cache.Config{CacheSize: 0, MaxEntry: 0, CacheMedia: media})
 }
 
-func testCacheGetPutHostHelper(t *testing.T, config *p2p.CacheConfig) {
+func testCacheGetPutHostHelper(t *testing.T, config *cache.Config) {
 	t.Helper()
 	Assert := assert.New(t)
-	c := p2p.NewCachePool(config)
+	c := cache.NewCachePool(config)
 	for i := 0; i < 1000; i++ {
 		wg.Add(1)
 		go func() {
 			defer wg.Done()
-			filename := p2p.GetRandomString(10)
-			host := p2p.GetRandomString(1024)
+			filename := util.GetRandomString(10)
+			host := util.GetRandomString(1024)
 			// get
 			res, hit := c.GetHost(filename)
 			Assert.Equal(false, hit)
@@ -95,18 +98,18 @@ func testCacheGetPutHostHelper(t *testing.T, config *p2p.CacheConfig) {
 }
 
 func TestCacheGetPutHost(t *testing.T) {
-	testCacheGetPutHostHelper(t, &p2p.CacheConfig{CacheSize: 0, MaxEntry: 1000 * 1024 * 1024, CacheMedia: media})
+	testCacheGetPutHostHelper(t, &cache.Config{CacheSize: 0, MaxEntry: 1000 * 1024 * 1024, CacheMedia: media})
 }
 
-func testCacheGetPutLengthHelper(t *testing.T, config *p2p.CacheConfig) {
+func testCacheGetPutLengthHelper(t *testing.T, config *cache.Config) {
 	t.Helper()
 	Assert := assert.New(t)
-	c := p2p.NewCachePool(config)
+	c := cache.NewCachePool(config)
 	for i := 0; i < 1000; i++ {
 		wg.Add(1)
 		go func() {
 			defer wg.Done()
-			filename := p2p.GetRandomString(10)
+			filename := util.GetRandomString(10)
 			length := rand.Int63n(1024)
 			// get
 			res, hit := c.GetLen(filename)
@@ -125,5 +128,5 @@ func testCacheGetPutLengthHelper(t *testing.T, config *p2p.CacheConfig) {
 }
 
 func TestCacheGetPutLength(t *testing.T) {
-	testCacheGetPutLengthHelper(t, &p2p.CacheConfig{CacheSize: 0, MaxEntry: 1000 * 1024 * 1024, CacheMedia: media})
+	testCacheGetPutLengthHelper(t, &cache.Config{CacheSize: 0, MaxEntry: 1000 * 1024 * 1024, CacheMedia: media})
 }
